hedgefund_simulator/requirements.txt
python-dotenv>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
yfinance>=0.2.3
scikit-learn>=1.0.0
matplotlib>=3.4.0
openai>=0.27.0
pytest>=7.0.0
